only preprocessing train set

+++ helpers/data_transforms.py
@@ -73,7 +73,7 @@ def local_phi_transformation(X, layers, phi, collection, phi_sector_index=None, 
 
 
     
-def load_in_data(collection_list, features, working_dir, training_frac, num_cond_features=0, feature_order=None, num_files=2, use_local_phi=False):
+def load_in_data(collection_list, features, working_dir, training_frac, num_cond_features=0, feature_order=None, num_files=1, use_local_phi=False):
         
     X = []
     layers = []
@@ -205,7 +205,7 @@ def logit_transform(x, all_min, all_max, cushion ):
 
 
 
-def preprocess_data(X, flow_training_dir, ZUKO_ID, num_cond_features=0):
+def preprocess_data(X, flow_training_dir, ZUKO_ID, num_cond_features=0, scaler_exists=False):
     """
     Preprocess data without modifying the original array.
     Applies log to the first column and standardizes all columns.
@@ -219,17 +219,29 @@ def preprocess_data(X, flow_training_dir, ZUKO_ID, num_cond_features=0):
         X_to_preproc = X
 
     if ZUKO_ID in ["UNAF", "NCSF"]:
-        min_max_scaler = MinMaxScaler(feature_range=(-np.pi, np.pi))
-        X_preproc = min_max_scaler.fit_transform(X_to_preproc)
-        with open(f"{flow_training_dir}/minmax", "wb") as ofile:
-            pickle.dump(min_max_scaler, ofile)
-
+        if scaler_exists:
+            with open(f"{flow_training_dir}/minmax", "rb") as ifile:
+                min_max_scaler = pickle.load(ifile)
+            X_preproc = min_max_scaler.transform(X_to_preproc)
+            
+        else:
+            min_max_scaler = MinMaxScaler(feature_range=(-np.pi, np.pi))
+            X_preproc = min_max_scaler.fit_transform(X_to_preproc)
+            with open(f"{flow_training_dir}/minmax", "wb") as ofile:
+                pickle.dump(min_max_scaler, ofile)
+    
 
     else:
-        standard_scaler = StandardScaler()
-        X_preproc = standard_scaler.fit_transform(X_to_preproc)
-        with open(f"{flow_training_dir}/standard", "wb") as ofile:
-            pickle.dump(standard_scaler, ofile)
+        if scaler_exists:
+            with open(f"{flow_training_dir}/standard", "rb") as ifile:
+                standard_scaler = pickle.load(ifile)
+            X_preproc = standard_scaler.inverse_transform(X_to_preproc)
+            
+        else:
+            standard_scaler = StandardScaler()
+            X_preproc = standard_scaler.fit_transform(X_to_preproc)
+            with open(f"{flow_training_dir}/standard", "wb") as ofile:
+                pickle.dump(standard_scaler, ofile)
 
 
     
@@ -250,7 +262,6 @@ def inverse_preprocess_data(X_preproc, flow_training_dir, ZUKO_ID, num_cond_feat
         X_to_unpreproc = X_preproc
 
         
-
 
     if ZUKO_ID in ["UNAF", "NCSF"]:
         with open(f"{flow_training_dir}/minmax", "rb") as ifile:
